수강신청 구현 완료

+++ src/main/java/com/green/muziuniv_be_notuser/app/student/enrollment/model/EnrollmentRes.java
@@ -11,7 +11,7 @@ import lombok.Setter;
 @AllArgsConstructor
 public class EnrollmentRes {
     private String status; //상태
-    private int courseId; //과목 아이디
+    private Long courseId; //과목 아이디
     private String courseCode; //과목 코드
     private String title; //강의 이름
     private String classroom; //강의실
